The End of a Journey
This is the last commit. What a journey this has been. The deadline is in an hour and 20 minutes. I haven't yet submitted the dissertation. What a truly incredible journey this has been.

I can now, finally, after all those days I dreamt of it, ride off into the sunset.

This project, and my time at university  as an undergrad has been spectacular